tasklog: make the shell integration tests pass
Since we're supressing the progress output when stdout is not a tty and the
shell integration tests run in a context where stdout is not a tty, all the
tests that assert content in the progress output are now failing. To alleviate
that, we run all commands for which we are asserting on their progress output in
a fake tty, utilizing the script(1) command.